Move Nuke.Common.IO to Nuke.Core.IO.
