+++ source/Nuke.Common/ChangeLog/ChangeLogTasks.cs
@@ -9,8 +9,8 @@ using System.IO;
 using System.Linq;
 using JetBrains.Annotations;
 using Nuke.Common.Git;
-using Nuke.Common.IO;
 using Nuke.Core;
+using Nuke.Core.IO;
 using Nuke.Core.Utilities;
 
 namespace Nuke.Common.ChangeLog
