+++ source/Nuke.Common/IconClasses.cs
@@ -8,7 +8,6 @@ using Nuke.Common;
 using Nuke.Common.ChangeLog;
 using Nuke.Common.Git;
 using Nuke.Common.Gitter;
-using Nuke.Common.IO;
 using Nuke.Common.Tools.CoverallsNet;
 using Nuke.Common.Tools.DocFx;
 using Nuke.Common.Tools.DotCover;
@@ -56,16 +55,8 @@ using Nuke.Core.Execution;
 [assembly: IconClass(typeof(OpenCoverTasks), "shield2")]
 [assembly: IconClass(typeof(PaketTasks), "box")]
 [assembly: IconClass(typeof(ReportGeneratorTasks), "pie-chart4")]
-[assembly: IconClass(typeof(SerializationTasks), "transmission2")]
 [assembly: IconClass(typeof(SignToolTasks), "key")]
 [assembly: IconClass(typeof(TestCloudTasks), "bug2")]
-[assembly: IconClass(typeof(TextTasks), "file-text3")]
 [assembly: IconClass(typeof(VsTestTasks), "bug2")]
 [assembly: IconClass(typeof(WebConfigTransformRunnerTasks), "bug2")]
-[assembly: IconClass(typeof(XmlTasks), "file-empty2")]
 [assembly: IconClass(typeof(XunitTasks), "bug2")]
-
-#if !NETCORE
-[assembly: IconClass(typeof(FtpTasks), "sphere2")]
-[assembly: IconClass(typeof(HttpTasks), "sphere2")]
-#endif
