+++ source/Nuke.Common/Tools/InspectCode/InspectCodeTasks.cs
@@ -7,7 +7,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using JetBrains.Annotations;
-using Nuke.Common.IO;
 using Nuke.Core;
 using Nuke.Core.BuildServers;
 using Nuke.Core.IO;
