+++ source/Nuke.Core/IO/HttpTasks.cs
@@ -6,10 +6,9 @@ using System;
 using System.Linq;
 using System.Net;
 using JetBrains.Annotations;
-using Nuke.Core.IO;
 using Nuke.Core.Tooling;
 
-namespace Nuke.Common.IO
+namespace Nuke.Core.IO
 {
     [PublicAPI]
     public static class HttpTasks
