+++ source/Nuke.Core/IconClasses.cs
@@ -17,3 +17,11 @@ using Nuke.Core.Tooling;
 [assembly: IconClass(typeof(ParameterAttribute), "syringe2")]
 [assembly: IconClass(typeof(PathConstruction), "price-tag2")]
 [assembly: IconClass(typeof(ProcessTasks), "terminal")]
+[assembly: IconClass(typeof(SerializationTasks), "transmission2")]
+[assembly: IconClass(typeof(TextTasks), "file-text3")]
+[assembly: IconClass(typeof(XmlTasks), "file-empty2")]
+
+#if !NETCORE
+[assembly: IconClass(typeof(FtpTasks), "sphere2")]
+[assembly: IconClass(typeof(HttpTasks), "sphere2")]
+#endif
