+++ source/Nuke.Core/NukeBuild.cs
@@ -8,10 +8,8 @@ using System.IO;
 using System.Linq;
 using System.Linq.Expressions;
 using JetBrains.Annotations;
-using Nuke.Core.BuildServers;
 using Nuke.Core.Execution;
 using Nuke.Core.IO;
-using Nuke.Core.OutputSinks;
 
 // ReSharper disable VirtualMemberNeverOverridden.Global
